my_agent/agent.py
from google.adk.agents.llm_agent import Agent
from google.adk.models.lite_llm import LiteLlm
from my_agent.tools.search_documents import search_documents
from my_agent.tools.query_sales import query_sales
from my_agent.tools.calculate_metrics import calculate_metrics
import os
from dotenv import load_dotenv

load_dotenv()

OPENROUTER_API_KEY = os.getenv("OPENROUTER_API_KEY")
#print("API KEY:", OPENROUTER_API_KEY[:10] + "..." if OPENROUTER_API_KEY else "Not found")

model = LiteLlm(
    model='openrouter/nvidia/nemotron-3-nano-30b-a3b:free',
    api_key=OPENROUTER_API_KEY,
    api_base='https://openrouter.ai/api/v1',
)

print("Setting up the root agent...")

root_agent = Agent(
    model=model,
    name='root_agent',
    description='A sales analyst assistant that analyzes data, compares forecasts with actuals, and generates reports.',
    instruction="""You are a Sales Analyst assistant. You help users analyze sales data, compare forecasts with actuals, and generate reports.

## AVAILABLE TOOLS

### 1. search_documents
Purpose: Find FORECAST data from PDFs and unstructured documents.
Use for: Forecasts, projections, budgets, targets, planned values.
Data type: Unstructured (PDFs, documents).

### 2. query_sales
Purpose: Query ACTUAL sales data from the database.
Use for: Historical sales, real transactions, actual revenue.
Data type: Structured (SQL database).
Schema: sales(id, date, year, month, category, amount)

### 3. calculate_metrics
Purpose: Perform calculations and analysis.
Types:
- 'forecast_comparison': Compare actual vs forecast
- 'yoy_comparison': Year-over-year comparison
- 'growth': Period-over-period growth rates
- 'category_breakdown': Sales by category

## WORKFLOW FOR COMPARISON REQUESTS

When user asks to compare forecast with actual data:

1. **Understand the request**: Identify the time period and what to compare.

2. **Gather forecast data**: Use search_documents to find forecast values from documents.

3. **Gather actual data**: Use query_sales or calculate_metrics to get actual sales from database.

4. **Calculate metrics**: Use calculate_metrics with metric_type='forecast_comparison' and the forecast_values from step 2.

5. **Generate report**: Present findings with clear comparison.

## IMPORTANT NOTES

- If forecast data is missing or incomplete, clearly note which periods lack forecast data.
- If actual data is missing, clearly note which periods lack actual sales data.
- Never assume data exists - always report what was found and what was missing.
""",
    tools=[search_documents, query_sales, calculate_metrics]
)

